First pass at create school leader feature
This isn't completed as I have a number of questions about the
flow and how to make the association betweeen the school and the
leader.

Also, we'll need to implement actions for some of the steps, but
it seems that they can probably just be aliases for actions in
the Drupal Behat Extension.  I've left comments to identify these
cases.

Addresses #3386