Refine school, user, profile features
Add extra fields to step to generate a school

Add building code and address fields to the definition of the
"Given the following schools exist:" step.
These fields show up via the relation to the School Profile
and are helpful for testing that the School Profile is properly
related to the School.

The address fields are implemented with
https://drupal.org/project/addressfield and are pretty nested.
The field names are also generalized to handle multiple localities.
So, in the semantics of this step, I've decide to use the names
reflected in the UI rather than the machine names of the fields.
I'll eventually have to translate between the field names when I
implement the step.

The step to create taxonomy terms that's included with the
Drupal Behat Extension doesn't handle extra fields on the
taxonomy entity.  We'll have to implement this ourselves.

http://drupalcode.org/project/addressfield.git/blob/refs/heads/7.x-1.x:/addressfield.devel_generate.inc
has examples of using the Field API to set the address subfields.

For the initial implementation, I might just ignore the address
fields and use the "batteries-included" step for creating the
taxonomy term.

Eventually, it might make more sense to use
https://drupal.org/project/migrate to create the taxonomy terms
en-masse.

Move the scenario for associating a school representative user
with profile to the school profile feature.

Addresses #3386